Add back nested folder useres/registrations_controller.rb
