+++ app/controllers/users/registrations_controller.rb
@@ -1,0 +1,18 @@
+# frozen_string_literal: true
+
+module Users
+  class RegistrationsController < Devise::RegistrationsController
+    before_action :configure_sign_up_params, only: [:create]
+    before_action :configure_account_update_params, only: [:update]
+
+  protected
+
+    def configure_sign_up_params
+      devise_parameter_sanitizer.permit(:sign_up, keys: %i[first_name last_name])
+    end
+
+    def configure_account_update_params
+      devise_parameter_sanitizer.permit(:account_update, keys: %i[first_name last_name])
+    end
+  end
+end
